Restructure code
Server can now be started in a normal binary by passing --server
A server binary can also be created with the `server-only` Cargo feature
