--- src/bin/server.rs
@@ -1,5 +1,0 @@
-extern crate space_shooter;
-
-fn main() {
-    space_shooter::Server::new().run()
-}
