+++ src/obj.rs
@@ -1,5 +1,5 @@
-use super::Vect;
-use simple_vector2d::Vector2;
+pub type Vect = Vector2<f32>;
+pub use simple_vector2d::Vector2;
 
 #[derive(Default, Debug, Copy, Clone)]
 #[derive(RustcEncodable, RustcDecodable)]
@@ -41,5 +41,24 @@ impl BasicObject {
             position: Vector2(x, y),
             velocity: Vector2(vx, vy),
         }
+    }
+}
+
+const W: f32 = 1200./2.;
+const H: f32 =  900./2.;
+
+/// Wraps `p` if out of bounds
+pub fn stay_in_bounds(p: &mut Vect) {
+    if p.0 < -W {
+        p.0 += 2. * W;
+    }
+    if p.0 > W {
+        p.0 -= 2. * W;
+    }
+    if p.1 < -H {
+        p.1 += 2. * H;
+    }
+    if p.1 > H {
+        p.1 -= 2. * H;
     }
 }
